% Auto-generated: do not edit by hand
\name{attioTable}

\alias{attioTable}

\title{AttioTable component}

\description{

}

\usage{
attioTable(id=NULL, cellClassName=NULL, className=NULL,
colHeaders=NULL, columnSorting=NULL, columns=NULL,
contextMenu=NULL, data=NULL, dropdownMenu=NULL,
filters=NULL, headerClassName=NULL, height=NULL,
licenseKey=NULL, multiColumnSorting=NULL, rowHeaders=NULL,
rowHeight=NULL, settings=NULL, stretchH=NULL,
themeName=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks}

\item{cellClassName}{Character. Custom CSS class for table cells}

\item{className}{Character. Custom CSS class for the table container}

\item{colHeaders}{Logical | numeric | character | named list | unnamed list. Show column headers}

\item{columnSorting}{Logical | numeric | character | named list | unnamed list. Enable column sorting}

\item{columns}{List of logical | numeric | character | named list | unnamed lists. Column configuration}

\item{contextMenu}{Logical | numeric | character | named list | unnamed list. Enable context menu}

\item{data}{List of list of logical | numeric | character | named list | unnamed listss. Data for the table}

\item{dropdownMenu}{Logical | numeric | character | named list | unnamed list. Enable dropdown menu}

\item{filters}{Logical | numeric | character | named list | unnamed list. Enable filters}

\item{headerClassName}{Character. Custom CSS class for headers}

\item{height}{Character | numeric. Table height}

\item{licenseKey}{Character. License key for Handsontable}

\item{multiColumnSorting}{Logical | numeric | character | named list | unnamed list. Enable multi-column sorting}

\item{rowHeaders}{Logical | numeric | character | named list | unnamed list. Show row headers}

\item{rowHeight}{Numeric. Custom row height}

\item{settings}{Logical | numeric | character | named list | unnamed list. Any additional Handsontable settings}

\item{stretchH}{Character. Stretch columns to fill container}

\item{themeName}{Character. Theme name - supports Handsontable native themes}

\item{width}{Character | numeric. Table width}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

